Preparation for moving datastore classes to module (#321)
